APIs/API_Authenticacao/requirements.txt
fastapi
uvicorn
cassandra-driver
pydantic
pyjwt
pika